fix(hw14): remove utils
