+++ hw12_13_14_15_calendar/pkg/rmq/rmq.go
@@ -1,8 +1,14 @@
 package rmq
 
 import (
+	"errors"
+	"fmt"
+
+	pkgnet "github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/pkg/net"
 	"github.com/streadway/amqp"
 )
+
+var ErrorUnsupportedProtocol = errors.New("invalid protocol")
 
 type MessageQueue struct {
 	Connection *amqp.Connection
@@ -33,4 +39,25 @@ func (m *MessageQueue) Close() error {
 		return err
 	}
 	return nil
+}
+
+func GetMqAddress(protocolArg, hostArg, portArg, userArg, passArg string) (string, error) {
+	var address string
+
+	if protocolArg != "amqp" {
+		return address, ErrorUnsupportedProtocol
+	}
+
+	netAddress, err := pkgnet.GetAddress(hostArg, portArg)
+	if err != nil {
+		return address, err
+	}
+
+	if userArg != "" && passArg != "" {
+		address = fmt.Sprintf("%v://%v:%v@%v", protocolArg, userArg, passArg, netAddress)
+	} else {
+		address = fmt.Sprintf("%v://%v", protocolArg, netAddress)
+	}
+
+	return address, nil
 }
